pyproject.toml
[tool.poetry]
name = "tarubot"
version = "0.1.0"
description = ""
authors = ["Connor Maddox <cmaddox@hack.foo>"]
readme = "README.md"

[tool.poetry.scripts]
start = "src.tarubot.main:main"

[tool.poetry.dependencies]
python = "^3.11"
python-box = "^7.1.1"
beautifulsoup4 = "^4.12.2"
coloredlogs = "^15.0.1"
dynaconf = "^3.2.4"
pony = "^0.7.17"
pymysql = "^1.1.0"
psycopg2-binary = "^2.9.9"
nextcord = "^2.6.0"
pynacl = "^1.5.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
